refactor(app): extract dashboard tile component
